Lesson_06 Homework

+++ Metrics/MetricsManager/Models/AgentInfo.cs
@@ -2,7 +2,7 @@
 {
     public class AgentInfo
     {
-        public int AgentId { get; set; }
+        public long AgentId { get; set; }
         public Uri AgentAddress { get; set; }
         public bool Enable { get; set; }
     }
